+++ integration-tests/global-setup/__tests__/setup3.test.js
@@ -1,0 +1,20 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+'use strict';
+
+const fs = require('fs');
+const os = require('os');
+const path = require('path');
+
+const DIR = path.join(os.tmpdir(), 'jest-global-setup');
+
+test('should exist setup file', () => {
+  const files = fs.readdirSync(DIR);
+  expect(files).toHaveLength(1);
+  const setup = fs.readFileSync(path.join(DIR, files[0]), 'utf8');
+  expect(setup).toBe('setup');
+});