+++ packages/jest-runner/src/test_worker.js
@@ -11,12 +11,7 @@ import type {GlobalConfig, Path, ProjectConfig} from 'types/Config';
 import type {SerializableError, TestResult} from 'types/TestResult';
 import type {RawModuleMap} from 'types/HasteMap';
 
-// Make sure uncaught errors are logged before we exit.
-process.on('uncaughtException', err => {
-  console.error(err.stack);
-  process.exit(1);
-});
-
+import exit from 'exit';
 import HasteMap from 'jest-haste-map';
 import {separateMessageFromStack} from 'jest-message-util';
 import Runtime from 'jest-runtime';
@@ -28,6 +23,12 @@ export type WorkerData = {|
   path: Path,
   rawModuleMap: ?RawModuleMap,
 |};
+
+// Make sure uncaught errors are logged before we exit.
+process.on('uncaughtException', err => {
+  console.error(err.stack);
+  exit(1);
+});
 
 const formatError = (error: string | Error): SerializableError => {
   if (typeof error === 'string') {